requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==1.10.13
requests==2.31.0
pdfplumber==0.10.3
openai==1.10.0
python-multipart==0.0.6
fpdf2==2.7.8

